toreja_pms_automation/requirements.txt
selenium>=4.18.0
pytest>=8.0.0
pytest-html>=4.1.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
